implement SIMDCAI av compression (it is called turbopfor because I was too lazy to change all the names)

+++ src/lib/storage/vector_compression/turboPFor_bitpacking/turboPFor_bitpacking_vector.cpp
@@ -2,10 +2,10 @@
 
 namespace opossum {
 
-TurboPForBitpackingVector::TurboPForBitpackingVector(pmr_vector<uint8_t>&& data, size_t size, uint8_t b) : _data{std::move(data)}, _size{size}, _b{b} {}
+TurboPForBitpackingVector::TurboPForBitpackingVector(pmr_vector<uint32_t>&& data, size_t size, uint8_t b) : _data{std::move(data)}, _size{size}, _b{b} {}
 
 size_t TurboPForBitpackingVector::on_size() const { return _size; }
-size_t TurboPForBitpackingVector::on_data_size() const { return sizeof(uint8_t) * _data.size(); }
+size_t TurboPForBitpackingVector::on_data_size() const { return sizeof(uint32_t) * _data.size(); }
 
 std::unique_ptr<BaseVectorDecompressor> TurboPForBitpackingVector::on_create_base_decompressor() const {
   return std::make_unique<TurboPForBitpackingDecompressor>(_data, _b, _size);
@@ -19,7 +19,7 @@ TurboPForBitpackingIterator TurboPForBitpackingVector::on_end() const { return T
 
 std::unique_ptr<const BaseCompressedVector> TurboPForBitpackingVector::on_copy_using_allocator(
     const PolymorphicAllocator<size_t>& alloc) const {
-  pmr_vector<uint8_t> data_copy(_data, alloc);
+  pmr_vector<uint32_t> data_copy(_data, alloc);
   return std::make_unique<TurboPForBitpackingVector>(std::move(data_copy), _size, _b);
 }
